Refactor purchaseMetaprice into its own tsx file
