+++ client/me/purchases/manage-purchase/purchase-meta.jsx
@@ -1,16 +1,10 @@
 import {
-	getPlan,
-	getProductFromSlug,
 	isConciergeSession,
-	isDIFMProduct,
 	isDomainRegistration,
 	isDomainTransfer,
-	isEmailMonthly,
 	isJetpackPlan,
 	isJetpackProduct,
 	JETPACK_LEGACY_PLANS,
-	TERM_BIENNIALLY,
-	TERM_MONTHLY,
 } from '@automattic/calypso-products';
 import { Card } from '@automattic/components';
 import { localizeUrl } from '@automattic/i18n-utils';
@@ -42,7 +36,6 @@ import {
 	isRenewable,
 	isWithinIntroductoryOfferPeriod,
 	isIntroductoryOfferFreeTrial,
-	getDIFMTieredPurchaseDetails,
 } from 'calypso/lib/purchases';
 import { CALYPSO_CONTACT, JETPACK_SUPPORT } from 'calypso/lib/url/support';
 import { getCurrentUser, getCurrentUserId } from 'calypso/state/current-user/selectors';
@@ -53,6 +46,7 @@ import { managePurchase } from '../paths';
 import { canEditPaymentDetails, isJetpackTemporarySitePurchase } from '../utils';
 import AutoRenewToggle from './auto-renew-toggle';
 import PaymentInfoBlock from './payment-info-block';
+import PurchaseMetaPrice from './purchase-meta-price';
 
 export default function PurchaseMeta( {
 	purchaseId = false,
@@ -222,114 +216,6 @@ function PurchaseMetaOwner( { owner } ) {
 			</span>
 		</li>
 	);
-}
-
-function PurchaseMetaPrice( { purchase } ) {
-	const translate = useTranslate();
-	const { productSlug, productDisplayPrice } = purchase;
-	const plan = getPlan( productSlug ) || getProductFromSlug( productSlug );
-
-	if ( isOneTimePurchase( purchase ) || isDomainTransfer( purchase ) ) {
-		if ( isDIFMProduct( purchase ) ) {
-			const difmTieredPurchaseDetails = getDIFMTieredPurchaseDetails( purchase );
-			if ( difmTieredPurchaseDetails && difmTieredPurchaseDetails.extraPageCount > 0 ) {
-				const {
-					extraPageCount,
-					formattedCostOfExtraPages: costOfExtraPages,
-					formattedOneTimeFee: oneTimeFee,
-				} = difmTieredPurchaseDetails;
-				return (
-					<div>
-						<div>
-							{ translate( 'Service: %(oneTimeFee)s (one-time)', {
-								args: {
-									oneTimeFee,
-								},
-							} ) }
-						</div>
-						<div>
-							{ translate(
-								'%(extraPageCount)d extra page: %(costOfExtraPages)s (one-time)',
-								'%(extraPageCount)d extra pages: %(costOfExtraPages)s (one-time)',
-								{
-									count: extraPageCount,
-									args: {
-										extraPageCount,
-										costOfExtraPages,
-									},
-								}
-							) }
-						</div>
-					</div>
-				);
-			}
-		}
-
-		// translators: displayPrice is the price of the purchase with localized currency (i.e. "C$10")
-		return translate( '{{displayPrice/}} {{period}}(one-time){{/period}}', {
-			components: {
-				displayPrice: (
-					<span
-						// eslint-disable-next-line react/no-danger
-						dangerouslySetInnerHTML={ { __html: productDisplayPrice } }
-					/>
-				),
-				period: <span className="manage-purchase__time-period" />,
-			},
-		} );
-	}
-
-	if ( isIncludedWithPlan( purchase ) ) {
-		return translate( 'Free with Plan' );
-	}
-
-	const getPeriod = () => {
-		if ( isEmailMonthly( purchase ) ) {
-			return translate( 'month' );
-		}
-
-		if ( plan && plan.term ) {
-			switch ( plan.term ) {
-				case TERM_BIENNIALLY:
-					return translate( 'two years' );
-				case TERM_MONTHLY:
-					return translate( 'month' );
-			}
-		}
-
-		if ( purchase.billPeriodLabel ) {
-			switch ( purchase.billPeriodLabel ) {
-				case 'per year':
-					return translate( 'year' );
-				case 'per month':
-					return translate( 'month' );
-				case 'per week':
-					return translate( 'week' );
-				case 'per day':
-					return translate( 'day' );
-			}
-		}
-
-		return translate( 'year' );
-	};
-
-	const getPriceLabel = ( period ) => {
-		//translators: displayPrice is the price of the purchase with localized currency (i.e. "C$10"), %(period)s is how long the plan is active (i.e. "year")
-		return translate( '{{displayPrice/}} {{period}}/ %(period)s{{/period}}', {
-			args: { period },
-			components: {
-				displayPrice: (
-					<span
-						// eslint-disable-next-line react/no-danger
-						dangerouslySetInnerHTML={ { __html: productDisplayPrice } }
-					/>
-				),
-				period: <span className="manage-purchase__time-period" />,
-			},
-		} );
-	};
-
-	return getPriceLabel( getPeriod() );
 }
 
 function PurchaseMetaIntroductoryOfferDetail( { purchase } ) {
